pyproject.toml
[tool.poetry]
name = "pkmodel-zxara"
version = "0.6.8"
description = "Pharmacokinetic modelling of drug delivery"

authors = [
"Anissa Alloula <anissa.alloula@dtc.ox.ac.uk>", 
"Marina Luchner <marina.luchner@dtc.ox.ac.uk>", 
"George Shillcock <george.shillcock@dtc.ox.ac.uk>", 
"Mia Whitefield <mia.whitefield@dtc.ox.ac.uk>", 
"Jacob Wright <jacob.wright@dtc.ox.ac.uk>"
]

license = "MIT"
readme = "README.md"
repository = "https://github.com/mwhitemfldm/zxara_PK"
documentation = "https://zxara-pk.readthedocs.io/en/latest/"
packages = [{include = "pkmodel_zxara"}]

[tool.poetry.dependencies]
python = "^3.8"
attrs = "22.1.0"
contourpy = "1.0.5"
cycler = "0.11.0"
fonttools = "4.37.4"
iniconfig = "1.1.1"
kiwisolver = "1.4.4"
matplotlib = "3.6.1"
numpy = "1.23.4"
packaging = "21.3"
pillow = "9.2.0"
pluggy = "1.0.0"
py = "1.11.0"
pyparsing = "3.0.9"
python-dateutil = "2.8.2"
scipy = "1.9.2"
six = "1.16.0"
tomli = "2.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
flake8 = "^5.0.4"
sphinx = "^5.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
